Fix missing imports
