+++ GUI/src/components/MetricAndPeriodOptions/ColoredCheckbox/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 
 interface CheckBoxButtonProps {
     label: string,
