+++ GUI/src/components/MetricAndPeriodOptions/Divider/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 
 const Divider: React.FC = () => {
     return <div style={{
