+++ GUI/src/components/MetricAndPeriodOptions/ToggleButton/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 
 interface ToggleButtonProps {
     label: string,
