+++ GUI/src/components/MetricAndPeriodOptions/index.tsx
@@ -1,47 +1,17 @@
-import React, { useEffect, useState } from 'react'
+import React, { useMemo, useState } from 'react'
 import Card from '../Card';
+import { periodOptions } from './data';
 import Divider from './Divider';
 import MetricOptionsGroup from './MetricOptionGroup';
 import SubOptionsGroup from './SubOptionsGroup';
-import ToggleButton from './ToggleButton';
-
-import { Option } from "./types";
+import { MetricOptionsState, Option } from "./types";
 
 
-
-const options: Option[] = [
-    { id: 'today', labelKey: 'today' },
-    { id: 'yesterday', labelKey: 'yesterday' },
-    { id: 'last30days', labelKey: 'last30days' },
-    { id: 'month', labelKey: 'month', additionalKey: 'monthpicker' },
-    { id: 'period', labelKey: 'customperiod', additionalKey: 'customperiodpicker' },
-];
-
-const metricOptions: Option[] = [
-    {
-        id: 'metric1',
-        labelKey: 'metric1',
-        subOptions: [
-            { id: 'option1', labelKey: 'option1', color: 'red' },
-            { id: 'option2', labelKey: 'option2', color: 'green' },
-            { id: 'option3', labelKey: 'option3', color: 'blue' },
-        ]
-    },
-    { id: 'metric2', labelKey: 'metric2' },
-    { id: 'metric3', labelKey: 'metric3' },
-    { id: 'metric4', labelKey: 'metric4' },
-];
-
-interface MetricOptionsState {
-    period: string,
-    metric: string,
-
-    start: string,
-    end: string,
-    options: string[],
+interface MetricOptionsProps {
+    metricOptions: Option[],
 }
 
-const MetricOptions: React.FC = () => {
+const MetricOptions: React.FC<MetricOptionsProps> = ({ metricOptions }) => {
     const [selection, setSelection] = useState<MetricOptionsState>({
         period: '',
         metric: '',
@@ -51,14 +21,15 @@ const MetricOptions: React.FC = () => {
         options: [],
     })
 
-    const subOptions = metricOptions.find((x) => x.id === selection.metric)?.subOptions ?? [];
+    const subOptions = useMemo(() =>
+        metricOptions.find((x) => x.id === selection.metric)?.subOptions ?? [],
+        [selection.metric]
+    );
 
-
-    console.log(selection);
     return (
         <Card>
             <MetricOptionsGroup
-                options={options}
+                options={periodOptions}
                 label='Period'
                 onChange={(period) => setSelection({ ...selection, period, })}
                 onDatePicked={(start, end) => setSelection({ ...selection, start, end, })}
@@ -83,3 +54,4 @@ const MetricOptions: React.FC = () => {
 }
 
 export default MetricOptions
+export type { Option }
